moving getters into their own files
