+++ pkg/api/configsync/v1beta1/reposync_types.go
@@ -79,16 +79,6 @@ type RepoSyncSpec struct {
 	Override *OverrideSpec `json:"override,omitempty"`
 }
 
-// GetOverride creates an override or returns an existing one
-// use it if you need to ensure that you are assigning
-// to an object, but not to test for nil (current existance)
-func (rs *RepoSyncSpec) GetOverride() *OverrideSpec {
-	if rs.Override == nil {
-		rs.Override = &OverrideSpec{}
-	}
-	return rs.Override
-}
-
 // RepoSyncStatus defines the observed state of a RepoSync.
 type RepoSyncStatus struct {
 	Status `json:",inline"`
