+++ pkg/api/configsync/v1beta1/resource_override.go
@@ -17,8 +17,6 @@ package v1beta1
 import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"kpt.dev/configsync/pkg/api/configsync"
-	"kpt.dev/configsync/pkg/client/restconfig"
 )
 
 // OverrideSpec allows to override the settings for a reconciler pod
@@ -90,20 +88,4 @@ type ContainerResourcesSpec struct {
 	// memoryLimit allows one to override the memory limit of a container
 	// +optional
 	MemoryLimit resource.Quantity `json:"memoryLimit,omitempty"`
-}
-
-// GetReconcileTimeout returns reconcile timeout in string, defaulting to 5m if empty
-func GetReconcileTimeout(d *metav1.Duration) string {
-	if d == nil || d.Duration == 0 {
-		return configsync.DefaultReconcileTimeout.String()
-	}
-	return d.Duration.String()
-}
-
-// GetAPIServerTimeout returns the API server timeout in string, defaulting to 5s if empty
-func GetAPIServerTimeout(d *metav1.Duration) string {
-	if d == nil || d.Duration == 0 {
-		return restconfig.DefaultTimeout.String()
-	}
-	return d.Duration.String()
 }
